preprocess-data/requirements.txt
PyMuPDF>=1.22.0  # For PDF to image conversion
Pillow>=10.0.0   # For image processing
boto3>=1.28.0    # For AWS services